Resolve "location tests"

Closes #34

See merge request mkaasm/iti0302-2024-backend!18

+++ src/main/java/ee/taltech/iti0302project/app/entity/location/LocationConditionEntity.java
@@ -8,14 +8,15 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter @Setter
+@Getter
+@Setter
 @Entity
 @Table(name = "location_conditions")
 public class LocationConditionEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
 
 }
