+++ src/main/java/ee/taltech/iti0302project/app/entity/location/LocationStatusEntity.java
@@ -8,13 +8,14 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter @Setter
+@Getter
+@Setter
 @Entity
 @Table(name = "location_statuses")
 public class LocationStatusEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
 }
